requirements.txt
fastapi~=0.116.1
uvicorn~=0.35.0
pydantic~=2.11.7
colorama~=0.4.6
pytest
sqlalchemy~=2.0.41
passlib~=1.7.4
python-jose~=3.5.0
requests
bcrypt==4.0.1
aiofiles